Fix incorrect display of std::uint8_t default values in the CLI help message

+++ src/hictk/include/hictk/tools/config.hpp
@@ -32,11 +32,11 @@ struct BalanceICEConfig {
   bool in_memory{false};
   bool symlink_to_weight{true};
   bool stdout_{false};
-  std::uint8_t zstd_compression_lvl{3};
+  std::int16_t zstd_compression_lvl{3};
   std::size_t threads{1};
   std::size_t chunk_size{10'000'000};
 
-  std::uint8_t verbosity{4};
+  std::int16_t verbosity{4};
   bool force{false};
 };
 
@@ -54,11 +54,11 @@ struct BalanceSCALEConfig {
   bool in_memory{false};
   bool symlink_to_weight{true};
   bool stdout_{false};
-  std::uint8_t zstd_compression_lvl{3};
+  std::int16_t zstd_compression_lvl{3};
   std::size_t threads{1};
   std::size_t chunk_size{10'000'000};
 
-  std::uint8_t verbosity{4};
+  std::int16_t verbosity{4};
   bool force{false};
 };
 
@@ -72,7 +72,7 @@ struct BalanceVCConfig {
   bool symlink_to_weight{true};
   bool stdout_{false};
 
-  std::uint8_t verbosity{4};
+  std::int16_t verbosity{4};
   bool force{false};
 };
 
@@ -95,7 +95,7 @@ struct ConvertConfig {
   std::size_t threads{2};
   std::size_t chunk_size{10'000'000};
 
-  std::uint8_t verbosity{4};
+  std::int16_t verbosity{4};
   bool force{false};
 };
 
@@ -118,7 +118,7 @@ struct DumpConfig {
   hic::MatrixType matrix_type{hic::MatrixType::observed};
   hic::MatrixUnit matrix_unit{hic::MatrixUnit::BP};
   std::uint32_t resolution{};
-  std::uint8_t verbosity{2};
+  std::int16_t verbosity{2};
   bool force{false};
 };
 
@@ -131,11 +131,11 @@ struct FixMcoolConfig {
   bool check_base_resolution{false};
 
   bool in_memory{false};
-  std::uint8_t zstd_compression_lvl{3};
+  std::int16_t zstd_compression_lvl{3};
   std::size_t chunk_size{10'000'000};
 
   std::size_t threads{1};
-  std::uint8_t verbosity{4};
+  std::int16_t verbosity{4};
   bool force{false};
 };
 
@@ -165,7 +165,7 @@ struct LoadConfig {
   std::size_t threads{2};
   std::uint32_t compression_lvl{9};
 
-  std::uint8_t verbosity{4};
+  std::int16_t verbosity{4};
   std::size_t batch_size{10'000'000};
 };
 
@@ -184,7 +184,7 @@ struct MergeConfig {
   std::string count_type{"int"};
 
   bool force{false};
-  std::uint8_t verbosity{4};
+  std::int16_t verbosity{4};
 };
 
 struct MetadataConfig {
@@ -194,7 +194,7 @@ struct MetadataConfig {
   bool include_file_path{false};
   bool recursive{false};
 
-  std::uint8_t verbosity{2};
+  std::int16_t verbosity{2};
 };
 
 struct RenameChromosomesConfig {
@@ -202,7 +202,7 @@ struct RenameChromosomesConfig {
   std::filesystem::path path_to_name_mappings{};
   bool add_chr_prefix{false};
   bool remove_chr_prefix{false};
-  std::uint8_t verbosity{4};
+  std::int16_t verbosity{4};
 };
 
 struct ValidateConfig {
@@ -212,7 +212,7 @@ struct ValidateConfig {
   bool include_file_path{true};
   bool exhaustive{true};
   bool quiet{false};
-  std::uint8_t verbosity{4};
+  std::int16_t verbosity{4};
 };
 
 struct ZoomifyConfig {
@@ -232,7 +232,7 @@ struct ZoomifyConfig {
   bool skip_all_vs_all_matrix{false};
 
   bool force{false};
-  std::uint8_t verbosity{4};
+  std::int16_t verbosity{4};
 };
 
 // clang-format off
